+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix qrscanner when rotating the device or resizing the window
 - macOS: create a universal build that runs natively on arm64 and amd64
 - Show fiat amount at the time of the transaction in transaction history
+- Add BTC Direct private trading desk information
 
 # 4.45.0
 - Bundle BitBox02 firmware version v9.21.0